fix: use _Atomic keyword for kernel code
https://en.cppreference.com/w/c/language/atomic.html

+++ user/user.h
@@ -22,7 +22,7 @@ int getpid(void);
 char* sbrk(int);
 int sleep(int);
 int uptime(void);
-uint getreadcount(void);
+uint64 getreadcount(void);
 
 // ulib.c
 int stat(const char*, struct stat*);
